Added widget saving
Fixed export methods
Routes rework

+++ backend/controller/exportMethods.py
@@ -1,4 +1,5 @@
-import utils.export.exportUtils as exportUtils
+from config.init_config import get_config
+import exportMethods.exportUtils as exportUtils
 
 #############################################################################
 # Export API Management
@@ -16,6 +17,12 @@ def get_export_methods():
 
 
 def post_export_method(data):
+    # Check if the creation of export methods is allowed
+    config = get_config()
+    creation_allowed = config["EXPORT_METHODS_CONFIG"]["creation"]
+    if not creation_allowed:
+        return "Export method creation is not allowed", 403
+
     try:
         return exportUtils.add_export_method(data), 200
     except Exception as e:
@@ -23,6 +30,12 @@ def post_export_method(data):
 
 
 def delete_export_method(exportMethodId):
+    # Check if the deletion of export methods is allowed
+    config = get_config()
+    deletion_allowed = config["EXPORT_METHODS_CONFIG"]["deletion"]
+    if not deletion_allowed:
+        return "Export method deletion is not allowed", 403
+
     try:
         return exportUtils.delete_export_method(exportMethodId), 200
     except Exception as e:
